chore(core): move loan scheme options v2
