+++ mobile-app/app/screens/AppNavigator/screens/Loans/components/LoanSchemeOptions.tsx
@@ -5,14 +5,14 @@ import {
 } from "@components/SkeletonLoader";
 import {
   ThemedIcon,
-  ThemedText,
-  ThemedTouchableOpacity,
-  ThemedView,
+  ThemedTextV2,
+  ThemedTouchableListItem,
+  ThemedViewV2,
 } from "@components/themed";
 import { LoanScheme } from "@defichain/whale-api-client/dist/api/loan";
 import { tailwind } from "@tailwind";
 import { translate } from "@translations";
-
+import { StyleProp, ViewStyle } from "react-native";
 import { NumericFormat as NumberFormat } from "react-number-format";
 
 export interface WalletLoanScheme extends LoanScheme {
@@ -23,80 +23,87 @@ interface LoanSchemeOptionsP {
   isLoading: boolean;
   loanSchemes: WalletLoanScheme[];
   selectedLoanScheme?: LoanScheme;
+  containerStyle?: StyleProp<ViewStyle>;
   onLoanSchemePress: (scheme: LoanScheme) => void;
 }
 
-export function LoanSchemeOptions(props: LoanSchemeOptionsP): JSX.Element {
+export function LoanSchemeOptionsV2(props: LoanSchemeOptionsP): JSX.Element {
   return (
-    <View style={tailwind("mb-1")} testID="loan_scheme_options">
+    <ThemedViewV2
+      style={[
+        tailwind("mt-8 mb-1 rounded-t-lg-v2 rounded-b-lg-v2"),
+        props.containerStyle,
+      ]}
+      testID="loan_scheme_options"
+      light={tailwind("bg-mono-light-v2-00")}
+      dark={tailwind("bg-mono-dark-v2-00")}
+    >
       {props.isLoading ? (
-        <View style={tailwind("p-4")}>
+        <View>
           <SkeletonLoader row={6} screen={SkeletonLoaderScreen.VaultSchemes} />
         </View>
       ) : (
-        <>
+        <View style={tailwind("px-5")}>
           {props.loanSchemes.map((scheme, index) => (
-            <ThemedTouchableOpacity
+            <ThemedTouchableListItem
               key={scheme.id}
-              light={tailwind("bg-white border-gray-300", {
-                "border-primary-500":
-                  props.selectedLoanScheme?.id === scheme.id,
-                "border-gray-50": scheme.disabled === true,
-              })}
-              dark={tailwind("bg-gray-800 border-gray-700", {
-                "border-darkprimary-500":
-                  props.selectedLoanScheme?.id === scheme.id,
-                "border-gray-900": scheme.disabled === true,
-              })}
-              style={tailwind(
-                "py-2 px-5 rounded-lg border flex flex-row items-center mb-1"
-              )}
+              isLast={index === props.loanSchemes.length - 1}
+              styleProps="py-5.5 flex flex-row items-center"
               onPress={() => props.onLoanSchemePress(scheme)}
               disabled={scheme.disabled}
               testID={`loan_scheme_option_${index}`}
             >
-              <ThemedView
-                light={tailwind("border-gray-500", {
-                  "border-primary-500 bg-primary-500":
+              <View style={tailwind("flex-1 flex-col")}>
+                <LoanSchemeOptionData
+                  label="Min. collateral ratio"
+                  value={scheme.minColRatio}
+                  testId={`min_col_ratio_value_${index}`}
+                  suffix="%"
+                  disabled={scheme.disabled}
+                />
+                <LoanSchemeOptionData
+                  label="Vault interest"
+                  value={scheme.interestRate}
+                  testId={`interest_rate_value_${index}`}
+                  suffix={`% ${translate(
+                    "components/LoanSchemeOptions",
+                    "APR"
+                  )}`}
+                  disabled={scheme.disabled}
+                  containerStyle="pt-1"
+                />
+              </View>
+
+              <ThemedViewV2
+                light={tailwind("border-mono-light-v2-700", {
+                  "border-green-v2 bg-green-v2":
                     props.selectedLoanScheme?.id === scheme.id,
-                  "border-gray-100 bg-gray-50": scheme.disabled === true,
+                  "opacity-30": scheme.disabled === true,
                 })}
-                dark={tailwind("border-gray-400", {
-                  "border-darkprimary-500 bg-darkprimary-500":
+                dark={tailwind("border-mono-dark-v2-700", {
+                  "border-green-v2 bg-green-v2":
                     props.selectedLoanScheme?.id === scheme.id,
-                  "border-gray-800 bg-gray-900": scheme.disabled === true,
+                  "opacity-30": scheme.disabled === true,
                 })}
-                style={tailwind("rounded-full border w-4 h-4 mr-7")}
+                style={tailwind(
+                  "rounded-full border-1.5 w-6 h-6 items-center justify-center"
+                )}
               >
                 {props.selectedLoanScheme?.id === scheme.id && (
                   <ThemedIcon
                     iconType="MaterialIcons"
                     name="check"
-                    size={14}
-                    light={tailwind("text-white")}
-                    dark={tailwind("text-black")}
+                    size={22}
+                    light={tailwind("text-mono-light-v2-00")}
+                    dark={tailwind("text-mono-dark-v2-00")}
                   />
                 )}
-              </ThemedView>
-              <LoanSchemeOptionData
-                label="Min. collateralization ratio"
-                value={scheme.minColRatio}
-                testId={`min_col_ratio_value_${index}`}
-                suffix="%"
-                disabled={scheme.disabled}
-              />
-              <LoanSchemeOptionData
-                label="Interest rate"
-                value={scheme.interestRate}
-                testId={`interest_rate_value_${index}`}
-                suffix={`% ${translate("components/LoanSchemeOptions", "APR")}`}
-                disabled={scheme.disabled}
-              />
-            </ThemedTouchableOpacity>
+              </ThemedViewV2>
+            </ThemedTouchableListItem>
           ))}
-        </>
+        </View>
       )}
-    </View>
+    </ThemedViewV2>
   );
 }
 
@@ -106,36 +113,37 @@ function LoanSchemeOptionData(props: {
   testId: string;
   suffix?: string;
   disabled?: boolean;
+  containerStyle?: string;
 }): JSX.Element {
   return (
-    <View style={tailwind("flex-1")}>
-      <ThemedText
-        light={tailwind("text-gray-400", {
-          "text-gray-300": props.disabled === true,
+    <View style={tailwind("flex-1 flex-row", props.containerStyle)}>
+      <ThemedTextV2
+        light={tailwind("text-mono-light-v2-700", {
+          "text-opacity-30": props.disabled === true,
         })}
-        dark={tailwind("text-gray-500", {
-          "text-gray-600": props.disabled === true,
+        dark={tailwind("text-mono-dark-v2-700", {
+          "text-opacity-30": props.disabled === true,
         })}
-        style={tailwind("text-xs")}
+        style={tailwind("text-sm font-normal-v2")}
       >
-        {translate("components/LoanSchemeOptions", props.label)}
-      </ThemedText>
+        {`${translate("components/LoanSchemeOptions", props.label)}:`}
+      </ThemedTextV2>
       <NumberFormat
         displayType="text"
         suffix={props.suffix}
         renderText={(value: string) => (
-          <ThemedText
-            light={tailwind("text-gray-900", {
-              "text-gray-300": props.disabled === true,
+          <ThemedTextV2
+            light={tailwind("text-mono-light-v2-900", {
+              "text-opacity-30": props.disabled === true,
             })}
-            dark={tailwind("text-gray-50", {
-              "text-gray-600": props.disabled === true,
+            dark={tailwind("text-mono-dark-v2-900", {
+              "text-opacity-30": props.disabled === true,
             })}
-            style={tailwind("text-sm font-medium")}
+            style={tailwind("text-sm font-semibold-v2 pl-1")}
             testID={props.testId}
           >
             {value}
-          </ThemedText>
+          </ThemedTextV2>
         )}
         thousandSeparator
         value={props.value}
